% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosm.R
\name{hosm}
\alias{hosm}
\title{Creates high order spatial matrix of the distance between locations}
\usage{
hosm(data)
}
\arguments{
\item{data}{dataframes from distances between locations}
}
\value{
A list the order and spatial weighting matrix of the distance between locations
}
\description{
Creates high order spatial matrix of the distance between locations
}
\examples{
hosm(simulation1)
hosm(simulation2)
hosm(simulation3)
hosm(simulation4)
hosm(simulation5)

}
\references{
Mubarak, F., Aslanargun, A., & Sıklar, I. (2022). GSTARIMA Model with Missing Value for Forecasting Gold Price. Indonesian Journal of Statistics and Its Applications, 6(1), 90–100. https://doi.org/10.29244/ijsa.v6i1p90-100

Mubarak, F., Aslanargun, A., & Sıklar, I. (2021). High order spatial weighting matrix using Google Trends. Int J Res Rev, 8(11), 388–396. https://doi.org/10.52403/ijrr.20211150

Mubarak, F., Aslanargun, A., & Sıklar, İ. (2022). Higher-order spatial classification using Google trends data during      covid-19. Adv. Appl. Stat., 78, 93–103. https://doi.org/10.17654/0972361722052
}
