% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation1.R
\docType{data}
\name{simulation1}
\alias{simulation1}
\title{Simulation 1 for High Order Spatial Matrix}
\format{
A data frame with 4 locations:
\describe{
\item{X}{Name of Location}
\item{X1}{1st Location}
\item{X2}{2nd Location}
\item{X3}{3rd Location}
\item{X4}{4th Location}
}
}
\usage{
simulation1
}
\description{
Simulation 1 for High Order Spatial Matrix
}
\examples{
data(simulation1)
}
\keyword{datasets}
