% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation2.R
\docType{data}
\name{simulation2}
\alias{simulation2}
\title{Simulation 2 for High Order Spatial Matrix}
\format{
A data frame with 5 locations:
\describe{
\item{Location}{Name of Location}
\item{'Amman (Jordan)}{'Amman City in Jordan}
\item{Abu Dhabi (United Arab Emirates)}{Abu Dhabi City in United Arab Emirates}
\item{Abuja (Nigeria)}{Abuja City in Nigeria}
\item{Accra (Ghana)}{Accra City in Ghana}
\item{Adamstown (Pitcairn)}{Adamstown City in Pitcairn}
}
}
\usage{
simulation2
}
\description{
Simulation 2 for High Order Spatial Matrix
}
\examples{
data(simulation2)
}
\keyword{datasets}
