% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation3.R
\docType{data}
\name{simulation3}
\alias{simulation3}
\title{Simulation 3 for High Order Spatial Matrix}
\format{
A data frame with 5 locations:
\describe{
\item{Location}{Name of Location}
\item{Yaren (Nauru)}{Yaren City in Nauru}
\item{Yerevan (Armenia)}{Yerevan City in Armenia}
\item{Zagreb (Croatia)}{Zagreb City in Croatia}
\item{al-'Ayun (Western Sahara)}{al-'Ayun City in Western Sahara}
\item{al-Kuwayt (Kuwait)}{al-Kuwayt in (Kuwait)}
}
}
\usage{
simulation3
}
\description{
Simulation 3 for High Order Spatial Matrix
}
\examples{
data(simulation3)
}
\keyword{datasets}
