% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation4.R
\docType{data}
\name{simulation4}
\alias{simulation4}
\title{Simulation 4 for High Order Spatial Matrix}
\format{
A data frame with 4 locations:
\describe{
\item{Location}{Name of Location}
\item{Ankara (Turkey)}{Ankara City in Turkey}
\item{Jakarta (Indonesia)}{Jakarta City in Indonesia}
\item{London (UK)}{London City in UK}
\item{Washington (USA)}{Washington in USA}
}
}
\usage{
simulation4
}
\description{
Simulation 4 for High Order Spatial Matrix
}
\examples{
data(simulation4)
}
\keyword{datasets}
